feat(svg-annotation-overlay): add media SVG annotation overlay component
